fix-size_t

+++ ft_strmapi.c
@@ -19,7 +19,7 @@ char				*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 
 	if (s == NULL || f == NULL)
 		return (NULL);
-	str = malloc(sizeof(char) * (ft_strlen(s) + 1));
+	str = malloc(sizeof(char) * ((unsigned int)ft_strlen(s) + 1));
 	if (str == NULL)
 		return (NULL);
 	i = 0;
